Added new component for Wow-Fact slider
